Refactor classes to real names. Keep iOS compatibility.
