+++ Phoebe/Phoebe.Android.csproj
@@ -59,6 +59,7 @@
     <Compile Include="Threading\**\*.cs" />
     <Compile Include="Analytics\**\*.cs" />
     <Compile Include="Logging\**\*.cs" />
+    <Compile Include="Data\Views\AllTimeEntriesView.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Contrib\" />
